problem 07 ef core intro
this problem was quite interesting, because i needed me to use a nested anonymous object in order to obtain information about the specific projects and then was a bit tricky to put all the results 2gether. Lucily ternar operators came into place. Overall really happy with this problem it was the first one to actually need some effort. I really enjoyed it !