uploaded TaskBoardApp workshop
Implemented all its functionality working crud and some validation checks and error handling. Tho workshop was designed to have all its logic in controllers which is bad practice i tried dividing it into data and service layers thus making it quite more challenging because much of the validation checks and the use of user manager classes are permitted and only available in the web part of the project. Still i managed to draw most of the complex logic and create interfaces and methods inside them. This gave me much better understanding of terms of architecture and how really works the mvc model. Also got  a good handle of the basic functions of asp.net and identity framework. i enjoyed writing this. Time to dive into example exams and see for myself what this exam hides as challenges.