sheesh
actually really gotten into it. Especially into thinking for validation errors and stuff. I really like this way of programming. This is my 3rd exam and i am feeling pretty good. Can emphasize a bit more on comments for clarification of each method and making straightforward validations. All is good. I even find it enjoyable and amusing. Had some problems because i wanted to make the user to log in with username no with email and had some issues in the input model but i figured it out. :)