web Funtamentals Retake Exam - 21 December 2022
this exam was made to be solved with custom identity but i made small changes and made it work with default identity. Made a try to implement service layer and build a level of abstraction. Tried thinking of some error and validation handling. Decided and implemented a login with username and not a email, tho i need both of the for registration. It was quite fun. Cant wait to learn and become better at web programming.